Extracted GuiHandler into its own class and added a method to get all the installed upgrades as a Set<ItemStack>.
